requirements.txt
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
feedparser==6.0.10
idna==3.3
pycodestyle==2.9.0
PyMsgBox==1.0.9
pyperclip==1.8.2
python-dotenv==0.20.0
requests==2.28.1
sgmllib3k==1.0.0
toml==0.10.2
urllib3==1.26.11
